requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==4.1.4
django-cors-headers==3.14.0
django-grappelli==3.0.4
django-suit @ https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
